Implement fetching groups from the backend

+++ SharedExpensesApp/app/src/main/java/com/example/sharedexpensesapp/model/Debt.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 data class Debt(
     val debtorId: String,
     val creditorId: String,
-    val amount: Int
+    val amount: Double
 )
